check IMP 0.0

+++ ModelUtils/Estimator/Estimator.py
@@ -49,7 +49,7 @@ class Estimator(object):
                                 "The clipped ts")
         samples_scores = tf.Print(samples_scores, [tf.reduce_max(samples_scores), tf.reduce_min(samples_scores)],
                                   "The clipped ts")
-        checker = tf.reduce_sum(tf.cast(tf.greater_equal(tf.reduce_max(samples_scores, 1) - target_score, 0), tf.int32))
+        checker = tf.reduce_sum(tf.cast(tf.greater_equal(tf.reduce_max(samples_scores, 1) - target_score, 0), tf.float32))
         checker = tf.Print(checker,[checker], "The value of checker is")
         Z = tf.reduce_sum(tf.exp(samples_scores), 1) + 0*checker
         Z = tf.Print(Z, [tf.reduce_max(Z), tf.reduce_min(Z)], "The value of Z is")
